Split _Init from New across Agent/Output/BlinkLed; wire an interrupt-driven transition
Generalizes the split Agent already had (Agent_Init vs NewAgent) to
Output: Output_Init(out, pin, behave, delete, name, timeout, stack,
prio) does the GPIO setup and delegates to Agent_Init with whatever
behave/delete it's given, while NewOutput() is just the convenience
wrapper that calls it with Output's own defaults. This is what makes
extending Output worthwhile instead of extending Agent directly (as
Pwm/Input still do): a derived type can call Output_Init() itself,
passing its own behave/delete through, and reuse the actual hardware
setup rather than duplicating it.

BlinkLed now extends Output instead of Agent directly (Agent <- Output
<- BlinkLed, three levels, same first-member idiom all the way down):
BlinkLed_Behave()/Blinked_Delete() call SetOutputValue() instead of
gpio_put() directly - real behavioral reuse, not just structural.
Updated blinkled.h's docs, which still described the old direct-Agent
design.

Also fixed a format-string bug in Agent's PrintInfo() (%u for a
uint32_t, which is unsigned long on this target - same class of bug
already fixed once before in peripheral.c).

state.c: wired in2's GPIO interrupt to a state transition, as a worked
example of doing this safely. in2CB(), the actual interrupt-context
handler (called by Peripheral's dispatcher), only sets a flag -
interrupts should do as little as possible. StateLoop_Run(), running
in the state machine's own task, checks that flag once per cycle and
is the only place that calls SetState() - nothing about the transition
itself runs in interrupt context. StateError_Run() reaches this by
deleting "led" for real (DeleteAgent()), same as before.

peripheral.c: dropped agent1/in1/pwm1 for now (kept out1/led/in2),
which also happens to resolve the earlier out1/pwm1 pin conflict since
pwm1 is gone. agent/in1/pwm remain declared in Peripheral but unused -
flagged separately, not touched here.

+++ src/agent.c
@@ -91,8 +91,8 @@ Agent *NewAgent(char *name, int timeout, uint32_t uxStackDepth, UBaseType_t uxPr
 /** @copydoc PrintInfo */
 void PrintInfo(Agent *agent) {
     if (agent == NULL) return;
-    printf("\tNAME = %s\tTIMEOUT=%d\tSTACK=%u\tPRIO=%lu\n",
-           agent->name, agent->timeout, agent->uxStackDepth, (unsigned long)agent->uxPriority);
+    printf("\tNAME = %s\tTIMEOUT=%d\tSTACK=%lu\tPRIO=%lu\n",
+           agent->name, agent->timeout, (unsigned long)agent->uxStackDepth, (unsigned long)agent->uxPriority);
 }
 
 /** @copydoc Run */
